move dummy data from function to task
